Carousel bullshit
Can't find a decent javascript library for swiping with visual feedback. This one doesn't have a way of jumping to a slide without playing the slide animation so just hide the image until it's done. There might be a better way where we temporarily change the carousel options but this seemed a bit too fiddly